- Notification views refactored - ElasticSearch dependencies updated - ES search query corrected
