+++ instanotifier/api/urls.py
@@ -1,36 +1,30 @@
 from django.conf.urls import url, include
 from rest_framework.routers import DefaultRouter
 
-from instanotifier.api.notification import views as notification_views
+from instanotifier.api.notification.search import NotificationSearchViewSet
+from instanotifier.api.notification import endpoints
 
 app_name = "api-v1"
 
 router = DefaultRouter()
-router.include_root_view = False
 
 router.register(
     r'rss-search',
-    notification_views.NotificationSearchViewSet,
+    NotificationSearchViewSet,
     basename='rss'
+)
+router.register(
+    r'notifications',
+    endpoints.NotificationViewSet,
+    basename="notification"
 )
 
 urlpatterns = router.urls
 
 urlpatterns += [
     url(
-        r"^$",
-        notification_views.NotificationListView.as_view(),
-        name="rssnotification-list",
-    ),
-    # url(r'^notifications/(?P<pk>\d+)/rating/(?P<rating>[\w]+)$', notification_views.NotificationVotingView.as_view(), name='rssnotification-rating'),
-    url(
-        r"^notifications/rating/$",
-        notification_views.NotificationVotingView.as_view(),
-        name="rssnotification-rating",
-    ),
-    url(
         r"^dates/$",
-        notification_views.NotificationDatesListView.as_view(),
+        endpoints.NotificationDatesListEndpoint.as_view(),
         name="rssnotification-date-list",
     ),
 ]
