+++ requirements/base.txt
@@ -1,7 +1,7 @@
 ipython>=7.5.0
 
 celery==4.3.0
-Django==2.2.2
+Django==2.2.10
 djangorestframework==3.9.4
 coreapi==2.3.3
 
@@ -33,8 +33,7 @@ django-bootstrap-pagination==1.7.1
 
 stackprinter==0.2.3
 
-django-elasticsearch-dsl==7.0.0
-django-elasticsearch-dsl-drf==0.20.3
-elasticsearch==7.0.4
-elasticsearch-dsl==7.0.0
-
+elasticsearch==7.6.0
+elasticsearch-dsl==7.1.0
+django-elasticsearch-dsl==7.1.1
+django-elasticsearch-dsl-drf==0.20.8
